add client library version to registration message indicating that client supports (and uses) latest phantom selection algorithm
