+++ protobuf/test.go
@@ -1,0 +1,38 @@
+package tdproto
+
+import (
+	"testing"
+
+	"github.com/golang/protobuf/proto"
+	"github.com/stretchr/testify/require"
+)
+
+// Write a small go test using your APIMessage (serialize/deserialize)
+func TestBidirectionalAPIResponse(t *testing.T) {
+	c2s := RegistrationResponse{}
+	addr := uint32(12345)
+	c2s.Ipv4Addr = &addr
+	port := uint32(10)
+	c2s.Port = &port
+
+	// Serialize
+	marsh, err := proto.Marshal(&c2s)
+	require.Nil(t, err)
+
+	// Deserialize
+	deser := RegistrationResponse{}
+	err = proto.Unmarshal(marsh, &deser)
+	require.Nil(t, err)
+	require.Equal(t, addr, deser.GetIpv4Addr())
+	require.Equal(t, port, deser.GetPort())
+}
+
+// TestProtoLibVer validates that the accessor method returns a default value for
+// fields that are unset in a protobuf and that our initial incremented
+// ClientLibraryVersion should be 1.
+func TestProtoLibVer(t *testing.T) {
+	c2s := ClientToStation{}
+
+	defaultLibVer := c2s.GetClientLibVersion()
+	require.Equal(t, uint32(0), defaultLibVer)
+}
